services/builds/requirements.txt
# requirements.txt para services/builds
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic
pydantic-settings
httpx
python-dotenv
